Attempt to convert search to TypeScript
